+++ benchmark/matmul/matmul_registry.cpp
@@ -242,6 +242,11 @@ inline constexpr MatMulStridedLhsInterface kai_matmul_clamp_f16_f16_f16p2vlx2b_1
     .run_matmul = kai_run_matmul_clamp_f16_f16_f16p2vlx2b_1x16vl_qmx_dot,
 };
 
+inline constexpr MatMulUkernelApiInterface kai_matmul_clamp_f16_f16_f16p4vsx2bf16_1x32vs_sme2_dot_interface{
+    .get_config = [] { return kai_matmul_uker_config{}; },
+    .get_api = kai_matmul_clamp_f16_f16_f16p4vsx2bf16_1x32vs_sme2_dot,
+};
+
 // matmul_clamp_f16_f16p_f16p
 inline constexpr MatMulBaseInterface kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_sme2_mopa_interface{
     .run_matmul = kai_run_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_sme2_mopa,
@@ -253,6 +258,12 @@ inline constexpr MatMulBaseInterface kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2b_2
 
 inline constexpr MatMulBaseInterface kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_qmx_mopa_interface{
     .run_matmul = kai_run_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_qmx_mopa,
+};
+
+inline constexpr MatMulUkernelApiInterface kai_matmul_clamp_f16_f16p4vsx2_f16p4vsx2bf16_8vsx8vs_sme2_mopa_interface{
+    .get_config = [] { return kai_matmul_uker_config{}; },
+    .get_api = kai_matmul_clamp_f16_f16p4vsx2_f16p4vsx2bf16_8vsx8vs_sme2_mopa,
+    .flags = KAI_MATMUL_UKER_FLAGS_ARGS_CLAMP,
 };
 
 // matmul_clamp_f32_bf16p_bf16p
@@ -962,6 +973,10 @@ inline const std::array matmul_benchmarks{
         "kai_matmul_clamp_f16_f16_f16p16vsx2bf16_6x16vs_sve2p1_dot", kai_benchmark_matmul<MatMulUkernelApiInterface>,
         kai_matmul_clamp_f16_f16_f16p16vsx2bf16_6x16vs_sve2p1_dot_interface, DataType::FP16, MatMulOp::GEMM,
         test::cpu_check<test::cpu_has_sve2p1, test::cpu_has_fp16>),
+    RegisterBenchmark(
+        "kai_matmul_clamp_f16_f16_f16p4vsx2bf16_1x32vs_sme2_dot", kai_benchmark_matmul<MatMulUkernelApiInterface>,
+        kai_matmul_clamp_f16_f16_f16p4vsx2bf16_1x32vs_sme2_dot_interface, DataType::FP16, MatMulOp::GEMV,
+        test::cpu_has_sme2),
 
     // matmul_clamp_f16_f16p_f16p
     RegisterBenchmark(
@@ -976,6 +991,11 @@ inline const std::array matmul_benchmarks{
         "kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_qmx_mopa", kai_benchmark_matmul<MatMulBaseInterface>,
         kai_matmul_clamp_f16_f16p2vlx2_f16p2vlx2_2vlx2vl_qmx_mopa_interface, DataType::FP16, MatMulOp::GEMM,
         test::cpu_has_sme),
+    RegisterBenchmark(
+        "kai_matmul_clamp_f16_f16p4vsx2_f16p4vsx2bf16_8vsx8vs_sme2_mopa",
+        kai_benchmark_matmul<MatMulUkernelApiInterface>,
+        kai_matmul_clamp_f16_f16p4vsx2_f16p4vsx2bf16_8vsx8vs_sme2_mopa_interface, DataType::FP16, MatMulOp::GEMM,
+        test::cpu_has_sme2),
 
     // matmul_clamp_f32_bf16p_bf16p
     RegisterBenchmark(