+++ kai/ukernels/matmul/kai_matmul_pack_lhs.h
@@ -30,6 +30,24 @@ extern "C" {
 /// @return The micro-kernel API.
 struct kai_matmul_pack_lhs_uker_api kai_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme(void);
 
+/// Non-transposed LHS packing micro-kernel for 16-bit data.
+///
+/// Required CPU features:
+///   * FEAT_SME
+///
+/// Configuration parameters: none.
+///
+/// Operands:
+///   * lhs_packed - The packed LHS matrix.
+///     * LHS matrix: 16-bit data in 4vsx2 blocked format.
+///   * lhs - The LHS matrix.
+///     * LHS matrix: 16-bit data in plain format.
+///
+/// Supported flags: none.
+///
+/// @return The micro-kernel API.
+struct kai_matmul_pack_lhs_uker_api kai_matmul_pack_lhs_mxk_x16p4vsx2_x16_sme(void);
+
 /// Non-transposed LHS packing micro-kernel for 32-bit data.
 ///
 /// Required CPU features: