+++ test/nextgen/operators/matmul/pack_lhs/matmul_pack_lhs_wrapper_registry.hpp
@@ -23,6 +23,9 @@ namespace kai::test {
 /// Creates a wrapper for kai_lhs_pack_f32p2vlx1_f32_sme micro-kernel.
 [[nodiscard]] std::unique_ptr<KernelWrapper<MatShape>> create_matmul_lhs_pack_f32p2vlx1_f32_sme();
 
+/// Creates a wrapper for kai_matmul_pack_lhs_mxk_x16p4vsx2_x16_sme.
+[[nodiscard]] std::unique_ptr<KernelWrapper<MatShape>> create_matmul_pack_lhs_mxk_x16p4vsx2_x16_sme();
+
 /// Creates a wrapper for kai_matmul_pack_lhs_mxk_x32p4vsx1_x32_sme.
 [[nodiscard]] std::unique_ptr<KernelWrapper<MatShape>> create_matmul_pack_lhs_mxk_x32p4vsx1_x32_sme();
 
@@ -34,6 +37,10 @@ namespace kai::test {
 
 /// Checks if the portion produces non-empty LHS packing tiles for the x32p4vsx1 matmul operator.
 [[nodiscard]] bool is_shape_suitable_lhs_x32p4vsx1_x32_sme(
+    size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
+
+/// Checks if the portion produces non-empty LHS packing tiles for the x16p4vsx2 matmul operator.
+[[nodiscard]] bool is_shape_suitable_lhs_x16p4vsx2_x16_sme(
     size_t shape_m, size_t shape_n, size_t shape_k, const MatrixPortion& portion);
 
 /// Checks if the portion produces non-empty LHS packing tiles for the x8p4vsx4 matmul operator.